Update checkout action on workflows on ubuntu 20.04 (#5387)
- Only workflows running Ubuntu 20.04 or later can be updated as the
GLIBC that is needed for node 20+ can be updated now.
- Workflows that aren't updated are running Ubuntu 18.04 or older, those
will need to be moved to updated images shortly and will be updated
later in the original PR, #5021

Sample warning that is resolved:

```
Node.js 16 actions are deprecated. Please update the following actions to use Node.js 20: actions/checkout@v3. For more information see: https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/.
```